feat: 회원가입 페이지에 변경 반영
